Only return RDS database instances that are available
hiera-aws throws an exception when a new RDS database is being
provisioned, i.e. has a state different from "available", like
"modifying". In that state not all database attributes are available
causing some Ruby fetch operation to fail.

The fix is to let hiera-aws only return database instances that are in
state "available". This can be checked via the "db_instance_status"
attribute.